Add admin and hidden fields to member models
A Boolean field called "admin" has been added to the department_type model, defaulting to False. Moreover, the res_partner.py file was updated to include many-to-many fields for `department_hidden` and `department_admin`. These fields are computed based on the relation to the "department_ids" field. The order of department names is also now considered in the Department model.